fix: printf specifier mismatches at LOG_* call sites
Exposed by the new format(printf, 3, 4) attribute on the logger's log_write_impl, which
until now left every LOG_* call site in this library unchecked.

These are all specifier widenings where the argument is uint32_t, int32_t, TickType_t or
EventBits_t - each 32-bit on the xtensa ABI, so the printed values were already correct.
They are fixed anyway because they were the noise that hid genuine undefined behaviour
elsewhere in the same sweep: %f given an int in ESP32-MB8ART, and %s given a std::string
object in ESP32-MB8ART and ESP32-PersistentStorage.

+++ src/MQTTManager.cpp
@@ -811,7 +811,8 @@ void MQTTManager::reconnectTaskFunc(void* param) {
     while (manager->autoReconnect) {
         if (!manager->isConnected()) {
             // Wait for the reconnect delay FIRST
-            MQTTM_LOG_I("Waiting %d ms before reconnection attempt", manager->currentReconnectDelay);
+            MQTTM_LOG_I("Waiting %lu ms before reconnection attempt",
+                        static_cast<unsigned long>(manager->currentReconnectDelay));
             vTaskDelay(pdMS_TO_TICKS(manager->currentReconnectDelay));
             
             // Check again after delay in case we connected elsewhere
@@ -1013,9 +1014,9 @@ void MQTTManager::reconnectTimerCallback(TimerHandle_t xTimer) {
             return;
         }
 
-        MQTTM_LOG_I("Reconnection attempt %d/%d",
-            manager->reconnectAttempts + 1,
-            manager->reconnectConfig.maxAttempts);
+        MQTTM_LOG_I("Reconnection attempt %lu/%lu",
+            static_cast<unsigned long>(manager->reconnectAttempts + 1),
+            static_cast<unsigned long>(manager->reconnectConfig.maxAttempts));
 
         // Increment attempt counter BEFORE calling connect()
         // Note: connect() is async - it returns Ok when connection is initiated,
